Remove SplitDimensionM pass
